SSCS-10517 renamed new service to RpcVenueService and all associated changes
