+++ src/test/java/uk/gov/hmcts/reform/sscs/bulkscancore/handlers/CcdCallbackHandlerTest.java
@@ -49,7 +49,7 @@ import uk.gov.hmcts.reform.sscs.domain.SscsCaseDetails;
 import uk.gov.hmcts.reform.sscs.domain.transformation.SuccessfulTransformationResponse;
 import uk.gov.hmcts.reform.sscs.exceptions.InvalidExceptionRecordException;
 import uk.gov.hmcts.reform.sscs.helper.AppellantPostcodeHelper;
-import uk.gov.hmcts.reform.sscs.helper.RpcVenueHelper;
+import uk.gov.hmcts.reform.sscs.service.RpcVenueService;
 import uk.gov.hmcts.reform.sscs.helper.SscsDataHelper;
 import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
 import uk.gov.hmcts.reform.sscs.model.CourtVenue;
@@ -87,7 +87,7 @@ public class CcdCallbackHandlerTest {
     private RefDataService refDataService;
 
     @Mock
-    private RpcVenueHelper rpcVenueHelper;
+    private RpcVenueService rpcVenueService;
 
     private CcdCallbackHandler ccdCallbackHandler;
 
@@ -109,18 +109,16 @@ public class CcdCallbackHandlerTest {
         fooLogger.addAppender(listAppender);
 
         SscsDataHelper sscsDataHelper = new SscsDataHelper(new CaseEvent(null, "validAppealCreated", null, null), dwpAddressLookupService, airLookupService, appellantPostcodeHelper, true);
-        ccdCallbackHandler = new CcdCallbackHandler(caseTransformer, caseValidator, sscsDataHelper, dwpAddressLookupService, refDataService, rpcVenueHelper, true);
+        ccdCallbackHandler = new CcdCallbackHandler(caseTransformer, caseValidator, sscsDataHelper, dwpAddressLookupService, refDataService, rpcVenueService, true);
 
         idamTokens = IdamTokens.builder().idamOauth2Token(TEST_USER_AUTH_TOKEN).serviceAuthorization(TEST_SERVICE_AUTH_TOKEN).userId(TEST_USER_ID).build();
 
-        given(dwpAddressLookupService.getDwpRegionalCenterByBenefitTypeAndOffice("PIP", "3"))
-            .willReturn("Springburn");
-        given(dwpAddressLookupService.getDwpRegionalCenterByBenefitTypeAndOffice("ESA", "Balham DRT"))
-            .willReturn("Balham");
+        given(dwpAddressLookupService.getDwpRegionalCenterByBenefitTypeAndOffice("PIP", "3")).willReturn("Springburn");
+        given(dwpAddressLookupService.getDwpRegionalCenterByBenefitTypeAndOffice("ESA", "Balham DRT")).willReturn("Balham");
 
         given(airLookupService.lookupAirVenueNameByPostCode(anyString(), any(BenefitType.class))).willReturn(PROCESSING_VENUE);
         when(refDataService.getVenueRefData(PROCESSING_VENUE)).thenReturn(CourtVenue.builder().regionId(REGION_ID).venueName(PROCESSING_VENUE).postcode(VENUE_POSTCODE).build());
-        when(rpcVenueHelper.retrieveRpcEpimsIdForAppellant(any())).thenReturn(EPIMMS_ID);
+        when(rpcVenueService.retrieveRpcEpimsIdForAppellant(any())).thenReturn(EPIMMS_ID);
 
         LocalDate localDate = LocalDate.now();
 
@@ -274,7 +272,7 @@ public class CcdCallbackHandlerTest {
         CaseResponse caseValidationResponse = CaseResponse.builder().build();
         when(caseValidator.validateValidationRecord(any(), anyBoolean())).thenReturn(caseValidationResponse);
         when(appellantPostcodeHelper.resolvePostcode(appeal.getAppellant())).thenReturn("CV35 2TD");
-        when(rpcVenueHelper.retrieveRpcEpimsIdForAppellant(appeal.getAppellant())).thenReturn("rpcEpimsId");
+        when(rpcVenueService.retrieveRpcEpimsIdForAppellant(appeal.getAppellant())).thenReturn("rpcEpimsId");
 
         PreSubmitCallbackResponse<SscsCaseData> ccdCallbackResponse = invokeValidationCallbackHandler(caseDetails.getCaseData());
